Provide condition to check whether build dir exists
- Remove build dir only if it exists

+++ build.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 # Remove build folder
-rm -r build/
+if [[ -d "build" ]]; then
+    rm -r build/
+fi
 
 # Create clear build folder and go inside
 mkdir build && cd build
